+++ include/ignition/math/MassMatrix3.hh
@@ -23,6 +23,7 @@
 
 #include <ignition/math/config.hh>
 #include "ignition/math/Helpers.hh"
+#include "ignition/math/Material.hh"
 #include "ignition/math/Quaternion.hh"
 #include "ignition/math/Vector2.hh"
 #include "ignition/math/Vector3.hh"
@@ -180,42 +181,90 @@ namespace ignition
 
       /// \brief Get IXX
       /// \return IXX value
-      public: T IXX() const
+      /// \deprecated See T Ixx() const
+      public: T IGN_DEPRECATED(5.0) IXX() const
+      {
+        return this->Ixx();
+      }
+
+      /// \brief Get IXX
+      /// \return IXX value
+      public: T Ixx() const
       {
         return this->Ixxyyzz[0];
       }
 
       /// \brief Get IYY
       /// \return IYY value
-      public: T IYY() const
+      /// \deprecated See T Iyy() const
+      public: T IGN_DEPRECATED(5.0) IYY() const
+      {
+        return this->Iyy();
+      }
+
+      /// \brief Get IYY
+      /// \return IYY value
+      public: T Iyy() const
       {
         return this->Ixxyyzz[1];
       }
 
       /// \brief Get IZZ
       /// \return IZZ value
-      public: T IZZ() const
+      /// \deprecated See T Izz() const
+      public: T IGN_DEPRECATED(5.0) IZZ() const
+      {
+        return this->IZZ();
+      }
+
+      /// \brief Get IZZ
+      /// \return IZZ value
+      public: T Izz() const
       {
         return this->Ixxyyzz[2];
       }
 
       /// \brief Get IXY
       /// \return IXY value
-      public: T IXY() const
+      /// \deprecated See T Ixy() const
+      public: T IGN_DEPRECATED(5.0) IXY() const
+      {
+        return this->Ixy();
+      }
+
+      /// \brief Get IXY
+      /// \return IXY value
+      public: T Ixy() const
       {
         return this->Ixyxzyz[0];
       }
 
       /// \brief Get IXZ
       /// \return IXZ value
-      public: T IXZ() const
+      /// \deprecated See T Ixz() const
+      public: T IGN_DEPRECATED(5.0) IXZ() const
+      {
+        return this->Ixz();
+      }
+
+      /// \brief Get IXZ
+      /// \return IXZ value
+      public: T Ixz() const
       {
         return this->Ixyxzyz[1];
       }
 
       /// \brief Get IYZ
       /// \return IYZ value
-      public: T IYZ() const
+      /// \deprecated See T Iyz() const
+      public: T IGN_DEPRECATED(5.0) IYZ() const
+      {
+        return this->Iyz();
+      }
+
+      /// \brief Get IYZ
+      /// \return IYZ value
+      public: T Iyz() const
       {
         return this->Ixyxzyz[2];
       }
@@ -223,16 +272,16 @@ namespace ignition
       /// \brief Set IXX
       /// \param[in] _v IXX value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIXX(const T &_v)
+      /// \deprecated See bool SetIxx(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IXX(const T &_v)
       {
-        return this->SetIXX(_v);
+        return this->SetIxx(_v);
       }
 
       /// \brief Set IXX
       /// \param[in] _v IXX value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIXX(const T &_v)
+      public: bool SetIxx(const T &_v)
       {
         this->Ixxyyzz.X(_v);
         return this->IsValid();
@@ -241,16 +290,16 @@ namespace ignition
       /// \brief Set IYY
       /// \param[in] _v IYY value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIYY(const T &_v)
+      /// \deprecated see bool SetIyy(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IYY(const T &_v)
       {
-        return this->SetIYY(_v);
+        return this->SetIyy(_v);
       }
 
       /// \brief Set IYY
       /// \param[in] _v IYY value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIYY(const T &_v)
+      public: bool SetIyy(const T &_v)
       {
         this->Ixxyyzz.Y(_v);
         return this->IsValid();
@@ -259,16 +308,16 @@ namespace ignition
       /// \brief Set IZZ
       /// \param[in] _v IZZ value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIZZ(const T &_v)
+      /// \deprecated See bool SetIzz(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IZZ(const T &_v)
       {
-        return this->SetIZZ(_v);
+        return this->SetIzz(_v);
       }
 
       /// \brief Set IZZ
       /// \param[in] _v IZZ value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIZZ(const T &_v)
+      public: bool SetIzz(const T &_v)
       {
         this->Ixxyyzz.Z(_v);
         return this->IsValid();
@@ -277,16 +326,16 @@ namespace ignition
       /// \brief Set IXY
       /// \param[in] _v IXY value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIXY(const T &_v)
+      /// \deprecated See bool SetIxy(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IXY(const T &_v)
       {
-        return this->SetIXY(_v);
+        return this->SetIxy(_v);
       }
 
       /// \brief Set IXY
       /// \param[in] _v IXY value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIXY(const T &_v)
+      public: bool SetIxy(const T &_v)
       {
         this->Ixyxzyz.X(_v);
         return this->IsValid();
@@ -295,16 +344,16 @@ namespace ignition
       /// \brief Set IXZ
       /// \param[in] _v IXZ value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIXZ(const T &_v)
+      /// \deprecated See bool SetIxz(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IXZ(const T &_v)
       {
-        return this->SetIXZ(_v);
+        return this->SetIxz(_v);
       }
 
       /// \brief Set IXZ
       /// \param[in] _v IXZ value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIXZ(const T &_v)
+      public: bool SetIxz(const T &_v)
       {
         this->Ixyxzyz.Y(_v);
         return this->IsValid();
@@ -313,16 +362,16 @@ namespace ignition
       /// \brief Set IYZ
       /// \param[in] _v IYZ value
       /// \return True if the MassMatrix3 is valid.
-      /// \deprecated bool SetIXZ(const T &_v)
+      /// \deprecated See bool SetIyz(const T &_v)
       public: bool IGN_DEPRECATED(5.0) IYZ(const T &_v)
       {
-        return this->SetIYZ(_v);
+        return this->SetIyz(_v);
       }
 
       /// \brief Set IYZ
       /// \param[in] _v IYZ value
       /// \return True if the MassMatrix3 is valid.
-      public: bool SetIYZ(const T &_v)
+      public: bool SetIyz(const T &_v)
       {
         this->Ixyxzyz.Z(_v);
         return this->IsValid();
@@ -330,7 +379,7 @@ namespace ignition
 
       /// \brief returns Moments of Inertia as a Matrix3
       /// \return Moments of Inertia as a Matrix3
-      /// \deprecated Matrix3<T> Moi() const
+      /// \deprecated See Matrix3<T> Moi() const
       public: Matrix3<T> IGN_DEPRECATED(5.0) MOI() const
       {
         return this->Moi();
@@ -411,8 +460,8 @@ namespace ignition
         // Check if mass and determinants of all upper left submatrices
         // of moment of inertia matrix are positive
         return (this->mass > 0) &&
-               (this->IXX() > 0) &&
-               (this->IXX()*this->IYY() - std::pow(this->IXY(), 2) > 0) &&
+               (this->Ixx() > 0) &&
+               (this->Ixx()*this->Iyy() - std::pow(this->Ixy(), 2) > 0) &&
                (this->Moi().Determinant() > 0);
       }
 
@@ -853,6 +902,20 @@ namespace ignition
 
         return true;
       }
+
+      /// \brief Set inertial properties based on a Material and equivalent box.
+      /// \param[in] _mat Material that specifies a density. Uniform density
+      /// is used.
+      /// \param[in] _size Size of equivalent box.
+      /// \param[in] _rot Rotational offset of equivalent box.
+      /// \return True if inertial properties were set successfully.
+      public: bool SetFromBox(const Material &_mat,
+                              const Vector3<T> &_size,
+                            const Quaternion<T> &_rot = Quaternion<T>::Identity)
+     {
+        double volume = _size.X() * _size.Y() * _size.Z();
+        return this->SetFromBox(_mat.Density() * volume, _size, _rot);
+     }
 
       /// \brief Set inertial properties based on mass and equivalent box.
       /// \param[in] _mass Mass to set.